Update vvi_from_sf.R
The slot that points to the C++ representation of a spatRaster object has been renamed twice in the last year in the terra package (from ´@ptr´ to ´@pnt´ to ´@cpp´). It is therefor better to not explicitly refer to the slot by its name. This PR fixes the issue for the `GVI::vvi_from_sf` function.